Read Columbus gps logger .csv file

read_columbus() imports latitude and longitude from a Columbus gps logger file.

read_columbus(
  path,
  gps_format = "nmea",
  filter_fix = NULL,
  tzone_gps = "UTC",
  tzone_out = "America/New_York",
  participant_id = NULL,
  sample_col = NULL,
  sample_id = NULL
)

Arguments

path

string; filepath of input .csv

gps_format

string; format of the input file ('nmea' or 'csv')

filter_fix

logical; should coordinates with invalid GPS fix be filtered out? Default = NULL 'gps_format' must be set to 'NMEA.'

tzone_gps

character; time zone of the Columbus logger. Default = 'UTC'

tzone_out

character; time zone of the output file. Default = 'America/New_York.'

participant_id

user defined string to denote a personal identifier. This is useful if the GPS unit is deployed during personal sampling. If specified, a new column is created ('ID'). Default is NULL.

sample_col

character; user defined character string specifying the name of the column to denote sample ID. Default is NULL.

sample_id

user defined string to denote sample ID. If assigned, a value must also be supplied to sample_col. Default is NULL.

Value

a tibble.

Examples

if (FALSE) {

read_columbus(path, tzone_gps = "UTC", tzone_out = "America/New_York.")
}